+++ src/combo-chart/index.js
@@ -1,0 +1,1 @@
+export { RowanComboChart, createParetoData } from "./combo-chart.js";